Remove reserve mana system + Adjust armour effects
The reserve mana system would need a complete rebalencing of all spells to make sense, and it is implemented with a lot of really invasive mixins, making it hard to maintain. This new version is more focused on balencing armor and high-rarity spells, which requires much less mixins and still achieves the goal of nerfing spells in an interesting way.